refactor: rename ext module to contrib
