+++ tests/contrib/aws/test_routes.py
@@ -1,6 +1,6 @@
-from loafer.ext.aws.message_translators import SNSMessageTranslator, SQSMessageTranslator
-from loafer.ext.aws.providers import SQSProvider
-from loafer.ext.aws.routes import SNSQueueRoute, SQSRoute
+from loafer.contrib.aws.message_translators import SNSMessageTranslator, SQSMessageTranslator
+from loafer.contrib.aws.providers import SQSProvider
+from loafer.contrib.aws.routes import SNSQueueRoute, SQSRoute
 
 
 def test_sqs_route(dummy_handler):
